requirements.txt
numpy
bs4
requests
scipy
matplotlib
pandas
streamlit
seaborn
pybase64
plotly.express


